Move geoclient-jni cli/util code to new projects.
Split gov.nyc.doitt.gis.geoclient.jni.test package into
two new projects:
:geoclient-utils:cli - move JniTest and make an application.
:geoclient-utils:jni-test: move TestConfig classes to library.

Note: per old comments in JniTest, there seems to be some issues when
not providing a file argument. Needs to investigate if the
InputStream resources are being opened/closed properly.

However, all existing tests pass and a full build succeeds.
