+++ geoclient-jni/build.gradle
@@ -4,6 +4,11 @@ plugins {
     id 'geoclient.test-conventions'
 }
 
+dependencies {
+    integrationTestImplementation(project(':geoclient-utils:jni-test'))
+    testFixturesImplementation(project(':geoclient-utils:jni-test'))
+}
+
 Provider<Directory> generatedResourcesDirectory = layout.buildDirectory.dir("generated-resources/main")
 File generatedResources = generatedResourcesDirectory.get().asFile
 def libPath = file('/opt/geosupport/current/lib')
@@ -12,10 +17,10 @@ sourceSets {
   main {
     output.dir(generatedResources, builtBy: 'syncJniLib')
   }
-// This is a workaround: see the comment for the syncJniTestConf task below.
-  integrationTest {
-    output.dir(builtBy: 'syncJniTestConf', layout.buildDirectory.dir('resources/integrationTest'))
-  }
+//// This is a workaround: see the comment for the syncJniTestConf task below.
+//  integrationTest {
+//    output.dir(builtBy: 'syncJniTestConf', layout.buildDirectory.dir('resources/integrationTest'))
+//  }
 }
 
 // Exclude from spotless (gradle plugin) license header check because of custom header.
@@ -23,23 +28,23 @@ spotless.java {
     targetExclude 'src/main/java/gov/nyc/doitt/gis/geoclient/jni/util/NativeLibraryLocator.java'
 }
 
-// This is a workaround for an apparent Gradle or JUnit issue:
-// After moving the gov.nyc.doitt.gis.geoclient.test.GeoclientStub from the
-// "main" SourceSet to "testFixtures", the JUnit ParameterizedTest
-// src/integrationTest/gov/nyc/doitt/gis/geoclient/jni/GeoclientJniIntegrationTest
-// began failing because it couldn't load '/jni-test.conf' using packaged
-// jar file syntax:
-// "file:/<path>/build/libs/geoclient-jni-2.0.1-beta.jar!/jni-test.conf"
-// Perhaps because debugging showed the testFixtures jar first in the
-// classpath?
-// If the classes in the "main" SourceSet package
-// gov.nyc.doitt.gis.geoclient.jni.test are moved to their own subproject,
-// it's possible this workaround can be removed.
-tasks.register('syncJniTestConf', Sync) {
-    dependsOn('processResources')
-    from layout.buildDirectory.dir('resources/main')
-    into layout.buildDirectory.dir('resources/integrationTest')
-}
+//// This is a workaround for an apparent Gradle or JUnit issue:
+//// After moving the gov.nyc.doitt.gis.geoclient.test.GeoclientStub from the
+//// "main" SourceSet to "testFixtures", the JUnit ParameterizedTest
+//// src/integrationTest/gov/nyc/doitt/gis/geoclient/jni/GeoclientJniIntegrationTest
+//// began failing because it couldn't load '/jni-test.conf' using packaged
+//// jar file syntax:
+//// "file:/<path>/build/libs/geoclient-jni-2.0.1-beta.jar!/jni-test.conf"
+//// Perhaps because debugging showed the testFixtures jar first in the
+//// classpath?
+//// If the classes in the "main" SourceSet package
+//// gov.nyc.doitt.gis.geoclient.jni.test are moved to their own subproject,
+//// it's possible this workaround can be removed.
+//tasks.register('syncJniTestConf', Sync) {
+//    dependsOn('processResources')
+//    from layout.buildDirectory.dir('resources/main')
+//    into layout.buildDirectory.dir('resources/integrationTest')
+//}
 
 // Must be configured lazily because 'from' and 'destinationDir' properties are
 // created dynamically by SharedLibraryBinarySpec
