+++ geoclient-utils/cli/build.gradle
@@ -1,0 +1,30 @@
+plugins {
+    id 'application'
+    id 'geoclient.java-conventions'
+}
+
+dependencies {
+    implementation project(':geoclient-utils:jni-test')
+    implementation project(':geoclient-jni')
+}
+
+application {
+    //applicationDefaultJvmArgs = ['-Xcheck:jni']
+    mainClass = 'gov.nyc.doitt.gis.geoclient.cli.JniTest'
+}
+
+//
+// To run the app with arguments to the main method use --args="...".
+//
+// Example: see the usage message.
+//
+// ./gradlew :geoclient-utils:cli:run --args="--help"
+//
+// Example: run from the project root and use the :geoclient-utils:jni-test
+// project's jni-test.conf file.
+//
+// ./gradlew :g-u:cli:run --args="--file=$(pwd)/geoclient-utils/jni-test/src/main/resources/jni-test.conf"
+//
+// See:
+// https://docs.gradle.org/current/userguide/application_plugin.html#application_plugin
+//
