+++ geoclient-utils/jni-test/build.gradle
@@ -1,0 +1,7 @@
+plugins {
+    id 'geoclient.library-conventions'
+}
+
+dependencies {
+}
+
