+++ geoclient-utils/jni-test/src/main/java/gov/nyc/doitt/gis/geoclient/jni/test/TestConfigurationException.java
