+++ geoclient-utils/jni-test/src/main/resources/jni-test.conf
